move List fields logic to a Trait (task #2191)
